CI: push the built container also to quay.io
[1] changed the Github action to push the fedora-bot image only to GHCR,
instead of Quay.io. While this looked sensible, this broke our
downstream release workflow. Specifically a private version of this bot
called centos-bot is using the fedora-bot container image as a base
image and builds on it. And it pulls the image from the Quay.io.

The [1] change silently broke our downstream workflow and we are still
using a F34 build of the fedora-bot container.

Push the built container to both, GHCR and Quay.io.

[1] https://github.com/osbuild/fedora-bot/commit/2d53757659c247140bdf389786de592f8b90bd79

Signed-off-by: Tomáš Hozza <thozza@redhat.com>