fedora_bot: make PR merge response parsing more robust
It seems that the response, when merging a MR in Fedora Pagure, may not
include the 'message' property. In such case, the bot will fail with a
traceback due to KeyError and won't log the received message.

Make the code more robust when accessing the response and print the whole
response in case it was not in an expected format.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>