fedora-bot: don't call sys.exit() on errors
The bot code is written with an expectation that if any errors happen
while handling an update of a specific package, an error is printed, but
the code continues handling the next component in the list. However, it
turns out that the msg_error() function, called by some functions,
called sys.exit(1), resulting in a hard failure where it should really
just continue with the next component instead.

Modify the function to raise a RuntimeError, which can be caught by the
caller, instead of exiting.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>