fedora-bot: handle 'error' status in PR test results
Extend the list of gracefully handled test statuses in a Pagure PR by
'error'. Previously, the code would raise an exception, which would
result in the following component PRs not being processed.

In addition, unify all messages printed by the bot to contain consistent
information about the PR that is being handled and also list the test
results in unexpected situations.

Signed-off-by: Tomáš Hozza <thozza@redhat.com>